Trait tectonic::engines::IoEventBackend

pub trait IoEventBackend {
    fn output_opened(&mut self, _name: &OsStr) { ... }
    fn stdout_opened(&mut self) { ... }
    fn output_closed(&mut self, _name: OsString, _digest: DigestData) { ... }
    fn input_opened(&mut self, _name: &OsStr, _origin: InputOrigin) { ... }
    fn input_not_available(&mut self, _name: &OsStr) { ... }
    fn input_closed(&mut self, _name: OsString, _digest: Option<DigestData>) { ... }
}

The IoEventBackend trait allows the program driving the TeX engines to track its input and output access patterns. The CLI program uses this information to intelligently decide when to rerun the TeX engine, to choose which files to actually save to disk, and to emit Makefile rules describing the dependency of the outputs on the inputs.

All of the trait methods have default implementations that do nothing.

Provided Methods

fn output_opened(&mut self, _name: &OsStr)

This function is called when a file is opened for output.

fn stdout_opened(&mut self)

This function is called when the wrapped "standard output" ("console", "terminal") stream is opened.

fn output_closed(&mut self, _name: OsString, _digest: DigestData)

This function is called when an output file is closed. The "digest" argument specifies the cryptographic digest of the data that were written. Note that this function takes ownership of the name and digest.

fn input_opened(&mut self, _name: &OsStr, _origin: InputOrigin)

This function is called when a file is opened for input.

fn input_not_available(&mut self, _name: &OsStr)

This function is called when the engine attempted to open a file of the specified name but it was not available.

fn input_closed(&mut self, _name: OsString, _digest: Option<DigestData>)

This function is called when an input file is closed. The "digest" argument specifies the cryptographic digest of the data that were read, if available. This digest is not always available, if the engine used seeks while reading the file. Note that this function takes ownership of the name and digest.