Module tectonic::digest

This module just have a few helpers to tidy up the computation of digests in various places.

Structs

DigestComputer
The SHA-256 hash algorithm with the SHA-256 initial hash value.

DigestData

Constants

DIGEST_NAME

Traits

Digest
The Digest trait specifies an interface common to digest functions, such as SHA-1 and the SHA-2 family of digest functions.

Functions

bytes_to_hex

create

hex_to_bytes