Module tectonic::engines

The Engines module provides access to the various processing backends used by Tectonic: bibtex, TeX, and xdvipdfmx. The API for each of these is defined in a sub-module with the corresponding name.

Due to the way Rust's visibility rules work, this module contains a substantial private API that defines the interface between Tectonic's Rust code and the C/C++ code that the backends are (currently) implemented in.

Reexports

pub use self::tex::TexEngine;

pub use self::xdvipdfmx::XdvipdfmxEngine;

pub use self::bibtex::BibtexEngine;

Modules

bibtex

tex

xdvipdfmx

Structs

NoopIoEventBackend
This struct implements the IoEventBackend trait but does nothing.

Traits

IoEventBackend
The IoEventBackend trait allows the program driving the TeX engines to track its input and output access patterns. The CLI program uses this information to intelligently decide when to rerun the TeX engine, to choose which files to actually save to disk, and to emit Makefile rules describing the dependency of the outputs on the inputs.