Module tectonic::engines

Reexports

pub use self::tex::TexEngine;

pub use self::xdvipdfmx::XdvipdfmxEngine;

pub use self::bibtex::BibtexEngine;

Modules

bibtex

io_api

kpse_api

md5_api

tex

xdvipdfmx

Structs

FileSummary
A summary of the I/O that happened on a file. We record its access pattern, the cryptographic digest of the file when it was last read, and the cryptographic digest of the file as it was last written.

Enums

AccessPattern
Different patterns with which files may have been accessed by the underlying engines. Once a file is marked as ReadThenWritten or WrittenThenRead, its pattern does not evolve further.