Struct tectonic::errors::Error

pub struct Error(_, _);

The Error type.

This struct is made of three things:

an ErrorKind which is used to determine the type of the error.

a backtrace, generated when the error is created.

an error chain, used for the implementation of Error::cause().

Methods

impl Error

fn from_kind(kind: ErrorKind) -> Error

Constructs an error from a kind, and generates a backtrace.

fn kind(&self) -> &ErrorKind

Returns the kind of the error.

fn iter(&self) -> ErrorChainIter

Iterates over the error chain.

fn backtrace(&self) -> Option<&Backtrace>

Returns the backtrace associated with this error.

impl Error

fn dump_uncolorized(&self)

Write the information contained in this object to standard error in a somewhat user-friendly form.

The error_chain crate provides a Display impl for its Error objects that ought to provide this functionality, but I have had enormous trouble being able to use it. So instead we emulate their code. This function is also paralleled by the implementation in status::termcolor::TermcolorStatusBackend, which adds the sugar of providing nice colorization if possible. This function should only be used if a StatusBackend is not yet available in the running program.