Module tectonic::io

Reexports

pub use self::filesystem::FilesystemIo;

pub use self::genuine_stdout::GenuineStdoutIo;

pub use self::memory::MemoryIo;

pub use self::stack::IoStack;

Modules

filesystem

genuine_stdout

itarbundle

local_cache

memory

stack

testing

zipbundle

Structs

InputHandle
Input handles are basically Read objects with a few extras. We don't require the standard io::Seek because we need to provide a dummy implementation for GZip streams, which we wouldn't be allowed to do because both the trait and the target struct are outside of our crate.

OutputHandle

Enums

InputOrigin
What kind of source an input file ultimately came from. We keep track of this in order to be able to emit Makefile-style dependencies for input files. Right now, we only provide enough options to achieve this goal; we could add more.

OpenResult

Traits

InputFeatures

IoProvider

Functions

try_open_file